replace web-related functionality with stub gatherstats command
This repository will contain only management commands. Remove
configuration required for hosting a web application replace it with a
single management command which can make use of the Django
infrastructure.

The management command is a stub command at the moment in order to
separate functionality and this large removal of code into separate
commits.