implement gatherstats management command
The stub implementation of gatherstats is removed and replaced with an
actual implementation which uses the plumbing created by previous
commits.

The management command takes either a URL to fetch or a local file path.
In both cases the endpoint is set automatically but can be overridden
from the command line if necessary.